Embed game into microsoft gamestatemanagement
