+++ Core/Entities/Characters/enemy/EnemyTypes/Chaser.cs
@@ -4,7 +4,7 @@ using ECS2022_23.Core.Entities.Characters.enemy.enemyBehavior;
 using ECS2022_23.Core.World;
 using Microsoft.Xna.Framework;
 
-namespace ECS2022_23.Core.Entities.Characters.enemy;
+namespace ECS2022_23.Core.Entities.Characters.enemy.EnemyTypes;
 
 public class Chaser : Enemy
 {
